Documents

+++ src/UABundle/Controller/ConsultantDocumentController.php
@@ -10,7 +10,8 @@ use FOS\RestBundle\Controller\Annotations\Put;
 use FOS\RestBundle\Controller\Annotations\View;
 use FOS\RestBundle\Controller\FOSRestController;
 use KernelBundle\Entity\DocumentTemplate;
-use KernelBundle\Entity\ConsultantDocumentTemplate;
+use Symfony\Component\Finder\Exception\AccessDeniedException;
+use UABundle\Entity\ConsultantDocumentTemplate;
 use KernelBundle\Entity\User;
 use Nelmio\ApiDocBundle\Annotation\ApiDoc;
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\ParamConverter;
@@ -84,7 +85,7 @@ class ConsultantDocumentController extends FOSRestController
 
     /**
      * Get all the consultant_documents from a template
-     * @param DocumentTemplate $template
+     * @param ConsultantDocumentTemplate $template
      * @return array
      *
      * @ApiDoc(
@@ -104,7 +105,7 @@ class ConsultantDocumentController extends FOSRestController
      * @ParamConverter("template", class="KernelBundle:DocumentTemplate")
      * @Get("/consultant_documents/template/{id}", requirements={"id" = "\d+"})
      */
-    public function getConsultantDocumentsByTemplateAction(DocumentTemplate $template)
+    public function getConsultantDocumentsByTemplateAction(ConsultantDocumentTemplate $template)
     {
 
         $consultant_documents = $this->getDoctrine()->getRepository("UABundle:ConsultantDocument")
@@ -231,7 +232,14 @@ class ConsultantDocumentController extends FOSRestController
         $form = $this->createForm(new ConsultantDocumentType(), $consultant_document);
         $form->handleRequest($request);
 
+        if (!$this->get('ua.project.rights_service')->userHasRights($this->getUser(), $consultant_document->getConsultant()->getProject())) {
+            throw new AccessDeniedException();
+        }
+
         if ($form->isValid()) {
+            $file = $consultant_document->getFile();
+            $fileName = $this->get('document_uploader')->upload($file);
+            $consultant_document->setFile($fileName);
             $em = $this->getDoctrine()->getManager();
             $em->persist($consultant_document);
             $em->flush();
@@ -276,6 +284,10 @@ class ConsultantDocumentController extends FOSRestController
         $form = $this->createForm(new ConsultantDocumentType(), $consultant_document);
         $form->handleRequest($request);
 
+        if (!$this->get('ua.project.rights_service')->userHasRights($this->getUser(), $consultant_document->getConsultant()->getProject())) {
+            throw new AccessDeniedException();
+        }
+
         if ($form->isValid()) {
             $em = $this->getDoctrine()->getManager();
 
@@ -315,6 +327,10 @@ class ConsultantDocumentController extends FOSRestController
      */
     public function deleteConsultantDocumentAction(ConsultantDocument $consultant_document)
     {
+        if (!$this->get('ua.project.rights_service')->userHasRights($this->getUser(), $consultant_document->getConsultant()->getProject())) {
+            throw new AccessDeniedException();
+        }
+
         $em = $this->getDoctrine()->getManager();
         $em->remove($consultant_document);
         $em->flush();
