+++ src/UABundle/Controller/DeliveryDocumentController.php
@@ -9,8 +9,8 @@ use FOS\RestBundle\Controller\Annotations\Delete;
 use FOS\RestBundle\Controller\Annotations\Put;
 use FOS\RestBundle\Controller\Annotations\View;
 use FOS\RestBundle\Controller\FOSRestController;
-use KernelBundle\Entity\DocumentTemplate;
-use KernelBundle\Entity\DeliveryDocumentTemplate;
+use Symfony\Component\Finder\Exception\AccessDeniedException;
+use UABundle\Entity\DeliveryDocumentTemplate;
 use KernelBundle\Entity\User;
 use Nelmio\ApiDocBundle\Annotation\ApiDoc;
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\ParamConverter;
@@ -84,7 +84,7 @@ class DeliveryDocumentController extends FOSRestController
 
     /**
      * Get all the delivery_documents from a template
-     * @param DocumentTemplate $template
+     * @param DeliveryDocumentTemplate $template
      * @return array
      *
      * @ApiDoc(
@@ -104,7 +104,7 @@ class DeliveryDocumentController extends FOSRestController
      * @ParamConverter("template", class="KernelBundle:DocumentTemplate")
      * @Get("/delivery_documents/template/{id}", requirements={"id" = "\d+"})
      */
-    public function getDeliveryDocumentsByTemplateAction(DocumentTemplate $template)
+    public function getDeliveryDocumentsByTemplateAction(DeliveryDocumentTemplate $template)
     {
 
         $delivery_documents = $this->getDoctrine()->getRepository("UABundle:DeliveryDocument")
@@ -231,7 +231,14 @@ class DeliveryDocumentController extends FOSRestController
         $form = $this->createForm(new DeliveryDocumentType(), $delivery_document);
         $form->handleRequest($request);
 
+        if (!$this->get('ua.project.rights_service')->userHasRights($this->getUser(), $delivery_document->getDelivery()->getTask()->getProject())) {
+            throw new AccessDeniedException();
+        }
+
         if ($form->isValid()) {
+            $file = $delivery_document->getFile();
+            $fileName = $this->get('document_uploader')->upload($file);
+            $delivery_document->setFile($fileName);
             $em = $this->getDoctrine()->getManager();
             $em->persist($delivery_document);
             $em->flush();
@@ -276,6 +283,10 @@ class DeliveryDocumentController extends FOSRestController
         $form = $this->createForm(new DeliveryDocumentType(), $delivery_document);
         $form->handleRequest($request);
 
+        if (!$this->get('ua.project.rights_service')->userHasRights($this->getUser(), $delivery_document->getDelivery()->getTask()->getProject())) {
+            throw new AccessDeniedException();
+        }
+
         if ($form->isValid()) {
             $em = $this->getDoctrine()->getManager();
 
@@ -315,6 +326,10 @@ class DeliveryDocumentController extends FOSRestController
      */
     public function deleteDeliveryDocumentAction(DeliveryDocument $delivery_document)
     {
+        if (!$this->get('ua.project.rights_service')->userHasRights($this->getUser(), $delivery_document->getDelivery()->getTask()->getProject())) {
+            throw new AccessDeniedException();
+        }
+
         $em = $this->getDoctrine()->getManager();
         $em->remove($delivery_document);
         $em->flush();
